+++ apps/common/utils/__init__.py
@@ -9,3 +9,4 @@ from .ipip import *
 from .crypto import *
 from .random import *
 from .jumpserver import *
+from .ip import *
